refactor : 리뷰 반영 - AbstractCircuitBreakerHealthIndicator 추출로 중복 제거
